Added documentation
common.tf_util:
- Moved `calculate_accuracy` fn to here

selfboosting.learn_selfboosting:
- Added documentation to `learn` fn
- Moved `calculate_accuracy` out

selfboosting/README.md
- created for sub-package local needs
